requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
spacy>=3.6.0
pdfplumber>=0.10.0
python-docx>=1.0.0
openpyxl>=3.1.0
xlsxwriter>=3.1.0
oracledb>=2.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.18
psutil>=5.9.0
